program frame completed and vtk output added
the MolSim.cpp has been completed and the XYZ writer has been replaced by a vtk writer. Some refactoring has been done as well

+++ src/planets/Gravity.cpp
@@ -5,19 +5,19 @@
 #include "../Particle.h"
 #include "../utils/ArrayUtils.h"
 
-void computeGravitySinglePlanet(int planetIdx, std::vector<Particle>& planets);
+void computeGravitySinglePlanet(long unsigned int planetIdx, std::vector<Particle>& planets);
 
 void computeGravity(std::vector<Particle>& planets) {
-    for (int planetIdx = 0; planetIdx < planets.size(); planetIdx++) {
+    for (long unsigned int planetIdx = 0; planetIdx < planets.size(); planetIdx++) {
         computeGravitySinglePlanet(planetIdx, planets);
     }  
 }
 
-void computeGravitySinglePlanet(int planetIdx, std::vector<Particle>& planets) {
+void computeGravitySinglePlanet(long unsigned int planetIdx, std::vector<Particle>& planets) {
     std::array<double, 3> newForce;
     Particle planet = planets[planetIdx];  
 
-    for (int otherPlanetIdx = 0; otherPlanetIdx < planets.size(); otherPlanetIdx++) {
+    for (long unsigned int otherPlanetIdx = 0; otherPlanetIdx < planets.size(); otherPlanetIdx++) {
         if (planetIdx == otherPlanetIdx) {
             continue;
         }
